requirements.txt
# Research Orchestration Agent - Requirements

# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# LangGraph for workflow orchestration
langgraph>=0.2.0

# OpenAI (latest version for web_search support)
openai>=1.54.0

# Pydantic for data validation
pydantic>=2.0.0

# Vector database for RAG - pgvector
pgvector>=0.2.0
psycopg2-binary>=2.9.0
langchain-postgres>=0.0.6

# Backup: ChromaDB (if needed)
# chromadb>=0.4.22
# sentence-transformers>=2.2.0
# langchain-chroma>=0.1.0

# Additional utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0
tiktoken>=0.5.0

# PDF and document generation
reportlab>=4.0.0
fpdf2>=2.7.0
